remove table v6 and references
