+++ package.json
@@ -43,7 +43,6 @@
     "lodash": "^4.17.14",
     "prop-types": "^15.7.2",
     "react-grid-system": "^4.4.6",
-    "react-table-v6": "^6.8.6",
     "react-tippy": "^1.4.0",
     "react-transition-group": "^4.3.0",
     "recompose": "^0.30.0",
