+++ src/Table/TablePagination/index.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,227 +17,121 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/** @jsxImportSource @emotion/react */
-
-import { css } from '@emotion/react';
-import { styled } from '../../ThemeProvider';
-import ceil from 'lodash/ceil';
-import floor from 'lodash/floor';
-import range from 'lodash/range';
-import PropTypes from 'prop-types';
-import React from 'react';
+import { css, useTheme } from '@emotion/react';
 import { Select } from '../../form/Select';
 import { POPUP_POSITIONS } from '../../form/Select/styledComponents';
-import { Icon } from '../../Icon';
-import { Typography } from '../../Typography';
-import useTheme from '../../utils/useTheme';
-
-export const TableActionBar = (props) => {
-  const { variant = 'label', color = 'grey', component = 'div' } = props;
-
-  return (
-    <Typography
-      {...props}
-      variant={variant}
-      color={color}
-      component={component}
-      css={css`
-        min-height: 32px;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        background: white;
-        padding-top: 8px;
-        padding-bottom: 8px;
-        padding-left: 8px;
-      `}
-    />
-  );
-};
-
-const Arrow: React.ComponentType<{ transform?: string; className?: string }> = ({
-  transform,
-  className,
-}) => (
-  <Icon
-    className={className}
-    width="6px"
-    height="6px"
-    name="chevron_right"
-    fill="grey"
-    transform={transform}
-  />
-);
-
-const DoubleArrow: React.ComponentType<{ transform?: string }> = ({ transform }) => (
-  <>
-    <Arrow transform={transform} />
-    <Arrow
-      css={css`
-        position: relative;
-        left: -3px;
-      `}
-      transform={transform}
-    />
-  </>
-);
-
-const A = styled('a')`
-  ${({ theme }) => theme.typography.data as any};
-  background-color: #fff;
-  border-radius: 50%;
-  cursor: pointer;
-  display: inline-block;
-  height: 24px;
-  line-height: 24px;
-  text-align: center;
-  width: 24px;
-  margin-right: 2px;
-
-  &:hover {
-    background-color: ${({ theme }) => theme.colors.primary_4};
-  }
-`;
-
-const PageControl = styled('div')`
-  display: flex;
-  align-items: center;
-  & a {
-    cursor: pointer;
-    display: inline-block;
-    width: 24px;
-    text-align: center;
-  }
-`;
+import { Arrow, DoubleArrow, PageButton, PageControl, TableActionBar } from './styled';
+import { ceil, floor, range } from 'lodash';
+import { Updater } from '@tanstack/react-table';
 
 // given 1 5 5 or 2 5 5, return [0,1,2,3,4]
-function getPagesAround(p, num, pages) {
-  const l = p - floor(num / 2);
-  const r = p + ceil(num / 2);
-  if (r > pages) {
-    return range(pages - num, pages);
-  }
-
-  if (l < 0) {
-    return range(0, num);
-  }
-  return range(l, r);
+function getPagesAround(pageIndex: number, rangeSize: number, pageCount: number) {
+  const prevPage = pageIndex - floor(rangeSize / 2);
+  const nextPage = pageIndex + ceil(rangeSize / 2);
+  return pageIndex === pageCount - 1 // currently on last page
+    ? range(pageCount - rangeSize, pageCount)
+    : pageIndex === 0 // currently on first page
+    ? range(0, rangeSize)
+    : range(prevPage, nextPage);
 }
 
-export function TablePagination(props) {
-  // page is zero indexed!
-  const { pages, page, showPageSizeOptions, pageSizeOptions, pageSize, onPageSizeChange } = props;
-
+export const TablePaginationV8 = ({
+  canNextPage,
+  canPreviousPage,
+  nextPage,
+  pageCount,
+  pageIndex,
+  pageSize,
+  pageSizeOptions = [5, 10, 20, 25, 50, 100],
+  previousPage,
+  resetPageIndex,
+  setPageIndex,
+  setPageSize,
+  showPageSizeOptions,
+}: {
+  canNextPage: boolean;
+  canPreviousPage: boolean;
+  nextPage: () => void;
+  pageCount: number;
+  pageIndex: number;
+  pageSize: number;
+  pageSizeOptions?: number[];
+  previousPage: () => void;
+  resetPageIndex: (defaultState?: boolean) => void;
+  setPageIndex: (updater: Updater<number>) => void;
+  setPageSize: (updater: Updater<number>) => void;
+  showPageSizeOptions?: boolean;
+}) => {
   const theme = useTheme();
-
   return (
-    <TableActionBar
-      css={css`
-        & svg {
-          box-sizing: content-box !important;
-        }
-        & a > svg {
-          position: relative;
-        }
-      `}
-    >
+    <TableActionBar showPageSizeOptions={showPageSizeOptions}>
       {showPageSizeOptions ? (
-        <div
-          css={css`
-            display: flex;
-            align-items: center;
-          `}
-        >
+        <PageControl>
           Show
-          <div
+          <Select
+            aria-label="Select page size"
             css={css`
-              transform: scale(0.8);
-              //use z-index to make popup always appear on the top of other components
               z-index: 10;
+              transform: scale(0.8);
+              & [role='button'] {
+                min-width: 70px;
+              }
             `}
-          >
-            <Select
-              css={css`
-                & [role='button'] {
-                  min-width: 70px;
-                }
-              `}
-              aria-label="Select page size"
-              options={pageSizeOptions.map((v) => ({ content: v.toString(), value: v }))}
-              onChange={onPageSizeChange}
-              value={pageSize}
-              popupPosition={POPUP_POSITIONS.UP}
-            />
-          </div>
+            onChange={(value: string) => {
+              setPageSize(parseInt(value));
+              resetPageIndex();
+            }}
+            options={pageSizeOptions.map((option: number) => ({
+              content: option,
+              value: option,
+            }))}
+            popupPosition={POPUP_POSITIONS.UP}
+            value={pageSize.toString()}
+          />
           rows
-        </div>
-      ) : (
-        <div />
-      )}
-
+        </PageControl>
+      ) : null}
       <PageControl>
-        <div>
-          <A
-            onClick={() => {
-              if (page === 0) return;
-              props.onPageChange(0);
-            }}
-          >
-            <DoubleArrow transform="rotate(180)" />
-          </A>
-          <A
-            onClick={() => {
-              if (page === 0) return;
-              props.onPageChange(page - 1);
-            }}
-          >
-            <Arrow transform="rotate(180)" />
-          </A>
-          {getPagesAround(page, 5, pages).map(
-            (p) =>
-              p > -1 &&
-              p < pages && (
-                <A
-                  key={p}
-                  onClick={() => props.onPageChange(p)}
-                  css={css`
-                    background-color: ${page === p ? theme.colors.secondary_4 : ''};
-                  `}
-                >
-                  {p + 1}
-                </A>
-              ),
-          )}
-          <A
-            onClick={() => {
-              if (page === pages - 1) return;
-              props.onPageChange(page + 1);
-            }}
-          >
-            <Arrow />
-          </A>
-          <A
-            onClick={() => {
-              if (page === pages - 1) return;
-              props.onPageChange(pages - 1);
-            }}
-          >
-            <DoubleArrow />
-          </A>
-        </div>
+        <PageButton onClick={() => setPageIndex(() => 0)}>
+          <DoubleArrow transform="rotate(180)" />
+        </PageButton>
+        <PageButton
+          onClick={() => {
+            if (canPreviousPage) {
+              previousPage();
+            }
+          }}
+        >
+          <Arrow transform="rotate(180)" />
+        </PageButton>
+        {getPagesAround(pageIndex, 5, pageCount).map(
+          (pageItem) =>
+            pageItem > -1 &&
+            pageItem < pageCount && (
+              <PageButton
+                css={css`
+                  background-color: ${pageIndex === pageItem ? theme.colors.secondary_4 : ''};
+                `}
+                key={pageItem}
+                onClick={() => setPageIndex(() => pageItem)}
+              >
+                {pageItem + 1}
+              </PageButton>
+            ),
+        )}
+        <PageButton
+          onClick={() => {
+            if (canNextPage) {
+              nextPage();
+            }
+          }}
+        >
+          <Arrow />
+        </PageButton>
+        <PageButton onClick={() => setPageIndex(() => pageCount - 1)}>
+          <DoubleArrow />
+        </PageButton>
       </PageControl>
     </TableActionBar>
   );
-}
-
-TablePagination.propTypes = {
-  /*
-   * check https://github.com/tannerlinsley/react-table/blob/v6/src/pagination.js
-   */
-  pages: PropTypes.number.isRequired,
-  page: PropTypes.number.isRequired,
-  showPageSizeOptions: PropTypes.bool,
-  pageSizeOptions: PropTypes.array,
-  onPageSizeChange: PropTypes.func,
-  onPageChange: PropTypes.func,
 };
