+++ src/Table/TableTabs/index.tsx
