--- src/Table/selectTable.md
@@ -1,3 +1,0 @@
-- This SelectTable component is the Table component that has been wrapped with a [selectTable](https://github.com/tannerlinsley/react-table/tree/v6/#selecttable) HOC, with some exension for styling.
-
-- Import: `import { SelectTable } from '../Table'`
