+++ src/index.tsx
@@ -93,14 +93,13 @@ export * from './Progress';
 export * from './SubMenu';
 export * from './SystemAlert';
 export * from './Table';
-export * from './TableV8';
-export * from './TableV8/SimpleTable';
-export * from './TableV8/TableFilters';
-export * from './TableV8/TablePagination';
-export * from './TableV8/TableRowSelection';
-export * from './TableV8/TableTabs';
-export * from './TableV8/styled';
-export * from './TableV8/types';
+export * from './Table/SimpleTable';
+export * from './Table/TableFilters';
+export * from './Table/TablePagination';
+export * from './Table/TableRowSelection';
+export * from './Table/TableTabs';
+export * from './Table/styled';
+export * from './Table/types';
 export * from './Tabs';
 export * from './Tag';
 export * from './theme/types';
